refactor: move all functions about z3 from utils to a separate file
